refactor(runtime): split the 1211-line dispatch match into per-domain handler modules
Pure code motion: Cmd::Data/Download/Local/MusicServer/ServerLibrary and the
search/AI/streaming/scrobble/transfer arms move to pub(super) handlers under
runtime/dispatch/, leaving dispatch.rs a thin dispatcher (1513 -> 519 lines).
Arm-by-arm equivalence against the pre-split file was verified in review;
strings, channel ops, error paths and arm order are unchanged.
